+++ src/main/java/com/redhat/demos/quarkusretailstore/inventory/api/InventoryDTO.java
@@ -1,6 +1,5 @@
 package com.redhat.demos.quarkusretailstore.inventory.api;
 
-import com.redhat.demos.quarkusretailstore.products.ProductMaster;
 import com.redhat.demos.quarkusretailstore.products.api.ProductMasterDTO;
 
 import java.math.BigDecimal;
@@ -10,9 +9,9 @@ public class InventoryDTO {
 
     final ProductMasterDTO productMaster;
 
-    final BigDecimal unitCost;
+    final Double unitCost;
 
-    final BigDecimal maxRetailPrice;
+    final Double maxRetailPrice;
 
     final int orderQuantity;
 
@@ -28,7 +27,7 @@ public class InventoryDTO {
 
     final int maximumQuantity;
 
-    public InventoryDTO(ProductMasterDTO productMaster, BigDecimal unitCost, BigDecimal maxRetailPrice, int orderQuantity, int inStockQuantity, int backOrderQuantity, LocalDateTime lastStockDate, LocalDateTime lastSaleDate, int minimumQuantity, int maximumQuantity) {
+    public InventoryDTO(ProductMasterDTO productMaster, Double unitCost, Double maxRetailPrice, int orderQuantity, int inStockQuantity, int backOrderQuantity, LocalDateTime lastStockDate, LocalDateTime lastSaleDate, int minimumQuantity, int maximumQuantity) {
         this.productMaster = productMaster;
         this.unitCost = unitCost;
         this.maxRetailPrice = maxRetailPrice;
@@ -99,11 +98,11 @@ public class InventoryDTO {
         return productMaster;
     }
 
-    public BigDecimal getUnitCost() {
+    public Double getUnitCost() {
         return unitCost;
     }
 
-    public BigDecimal getMaxRetailPrice() {
+    public Double getMaxRetailPrice() {
         return maxRetailPrice;
     }
